Revert resize_bilinear optimizations.
These cause the following self tests to fail.
//tensorflow/core/kernels:resize_area_op_test (tolerance issue)
//tensorflow/core/kernels:crop_and_resize_op_test
//tensorflow/core/kernels:conv_ops_test
//tensorflow/core/kernels:resize_bilinear_op_test (timeout)

+++ tensorflow/core/kernels/image_resizer_state.h
@@ -37,13 +37,12 @@ limitations under the License.
 
 namespace tensorflow {
 
-// CalculateResizeScale determines the scaling factor.
-// Need to use double to avoid round off errors for large x,y coordinates.
-inline double CalculateResizeScale(int64 in_size, int64 out_size,
+// CalculateResizeScale determines the float scaling factor.
+inline float CalculateResizeScale(int64 in_size, int64 out_size,
                                   bool align_corners) {
   return (align_corners && out_size > 1)
-             ? ((double)(in_size - 1) / (double)(out_size - 1))
-             : ((double)in_size / (double)(out_size));
+             ? (in_size - 1) / static_cast<float>(out_size - 1)
+             : in_size / static_cast<float>(out_size);
 }
 
 struct ImageResizerState {
@@ -122,8 +121,8 @@ struct ImageResizerState {
   int64 in_height;
   int64 in_width;
   int64 channels;
-  double height_scale;
-  double width_scale;
+  float height_scale;
+  float width_scale;
   Tensor* output = nullptr;
 
  private:
